create initVite method
